+++ src/lib/acme-spring/src/main/java/org/niis/xroad/common/acme/spring/provider/AcmeProfileIdConnection.java
@@ -24,7 +24,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.niis.xroad.common.acme.spring;
+package org.niis.xroad.common.acme.spring.provider;
 
 import lombok.extern.slf4j.Slf4j;
 import org.bouncycastle.asn1.x509.KeyUsage;
