+++ src/security-server/admin-service/application/src/main/java/org/niis/xroad/securityserver/restapi/config/AcmeBeanConfig.java
@@ -25,32 +25,18 @@
  */
 package org.niis.xroad.securityserver.restapi.config;
 
-import lombok.extern.slf4j.Slf4j;
-import org.apache.catalina.connector.Connector;
-import org.apache.coyote.http11.Http11NioProtocol;
 import org.niis.xroad.common.acme.AccountKeystorePasswordProvider;
 import org.niis.xroad.common.acme.AcmeConfig;
 import org.niis.xroad.common.acme.AcmeProperties;
 import org.niis.xroad.common.acme.AcmeService;
-import org.niis.xroad.common.properties.NodeProperties;
-import org.niis.xroad.common.properties.config.keys.AdminServiceConfigKeys;
-import org.niis.xroad.common.properties.spring.SpringConditionConfig;
-import org.niis.xroad.securityserver.restapi.scheduling.AcmeClientWorker;
-import org.niis.xroad.securityserver.restapi.scheduling.CertificateRenewalScheduler;
-import org.springframework.boot.tomcat.servlet.TomcatServletWebServerFactory;
-import org.springframework.boot.web.server.WebServerFactoryCustomizer;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Condition;
-import org.springframework.context.annotation.ConditionContext;
-import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.core.type.AnnotatedTypeMetadata;
-import org.springframework.scheduling.TaskScheduler;
 
-@Slf4j
+/**
+ * Wires the ACME client beans that are specific to the Security Server admin-service: they depend on this
+ * admin-service's own {@link AcmeProperties} and {@link AcmeConfig} bean instances (bound from
+ * {@code AdminServiceConfigKeys}), so they stay here rather than in acme-spring.
+ */
 @Configuration
 public class AcmeBeanConfig {
 
@@ -68,57 +54,6 @@ public class AcmeBeanConfig {
     public AcmeService acmeService(AcmeProperties acmeProperties, AcmeConfig acmeConfig,
                                    AccountKeystorePasswordProvider accountKeystorePasswordProvider) {
         return new AcmeService(acmeProperties, acmeConfig, accountKeystorePasswordProvider);
-    }
-
-    @Profile("nontest")
-    @Bean
-    public WebServerFactoryCustomizer<TomcatServletWebServerFactory> acmeChallengeCustomizer(AcmeConfig acmeConfig) {
-        if (acmeConfig.isAcmeChallengePortEnabled()) {
-            return factory -> {
-                var connector = new Connector(Http11NioProtocol.class.getName());
-                int acmeChallengePort = acmeConfig.getAcmeChallengePort();
-                connector.setScheme("http");
-                connector.setPort(acmeChallengePort);
-                log.info("ACME challenge port enabled, listening on port {}", acmeChallengePort);
-                factory.addAdditionalConnectors(connector);
-            };
-        } else {
-            log.info("ACME challenge port is disabled");
-            return factory -> {
-                // no-op
-            };
-        }
-    }
-
-    @Order(Ordered.LOWEST_PRECEDENCE - 99)
-    @Bean
-    @Conditional(IsAcmeCertRenewalJobsActive.class)
-    @Profile("!test")
-    CertificateRenewalScheduler certificateRenewalScheduler(AcmeClientWorker acmeClientWorker, TaskScheduler taskScheduler,
-                                                            AcmeConfig acmeConfig) {
-        CertificateRenewalScheduler scheduler = new CertificateRenewalScheduler(acmeClientWorker, acmeConfig, taskScheduler);
-        scheduler.init();
-        return scheduler;
-    }
-
-    @Slf4j
-    public static class IsAcmeCertRenewalJobsActive implements Condition {
-        @Override
-        public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
-            var config = SpringConditionConfig.resolve(context.getEnvironment(), AdminServiceConfigKeys.instance());
-            return schedulingEnabled(config.value(AdminServiceConfigKeys.ACME_RENEWAL_ACTIVE));
-        }
-
-        static boolean schedulingEnabled(boolean renewalActive) {
-            if (!renewalActive) {
-                log.info("ACME certificate renewal configured to be inactive, job auto-scheduling disabled");
-            }
-            if (NodeProperties.isSecondaryNode()) {
-                log.info("This is a secondary cluster node, ACME certificate renewal job auto-scheduling disabled");
-                return false;
-            }
-            return renewalActive;
-        }
     }
 
 }
