+++ src/security-server/admin-service/application/src/test/java/org/niis/xroad/securityserver/restapi/scheduling/AcmeClientWorkerTest.java
@@ -42,6 +42,7 @@ import org.junit.Test;
 import org.niis.xroad.common.acme.AcmeConfig;
 import org.niis.xroad.common.acme.AcmeKeyPurpose;
 import org.niis.xroad.common.acme.AcmeServiceException;
+import org.niis.xroad.common.acme.spring.scheduling.CertificateRenewalScheduler;
 import org.niis.xroad.common.managementrequest.ManagementRequestSender;
 import org.niis.xroad.securityserver.restapi.config.AbstractFacadeMockingTestContext;
 import org.niis.xroad.securityserver.restapi.util.CertificateTestUtils;
